android build now compiling, unsure of the changes to the windows side of things - but it's a start

+++ src/cloud/Environment/EnvironmentPeer.h
@@ -21,7 +21,7 @@ public:
 	}
 
 	RESULT PrintEnvironmentPeer() {
-		DEBUG_LINEOUT("User ID %d Environment ID %d", m_userID, m_environmentID);
+		DEBUG_LINEOUT("User ID %d Environment ID %d", (int)m_userID, (int)m_environmentID);
 		DEBUG_LINEOUT("SDP Offer: %s", m_strSDPOffer.c_str());
 		return R_PASS;
 	}
