+++ src/core/types/Message.h
@@ -61,8 +61,8 @@ public:
 	long GetReceiverUserID() { return m_header.receiverUserID; }
 
 	virtual RESULT PrintMessage() {
-		DEBUG_LINEOUT("Message from id %d to id %d at %s size:%d", 
-			m_header.senderUserID, m_header.receiverUserID, ctime(&(m_header.timestamp)), m_header.messageSize);
+		DEBUG_LINEOUT("Message from id %d to id %d at %s size:%d",
+					  (int)m_header.senderUserID, (int)m_header.receiverUserID, ctime(&(m_header.timestamp)), (int)m_header.messageSize);
 
 		return R_PASS;
 	}
